requirements.txt
blinker==1.8.2
boto3==1.34.115
botocore==1.34.115
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==24.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
requests==2.32.2
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.3
